+++ src/shared_lib.rs
@@ -16,9 +16,8 @@ use crate::os::windows::shared_lib::{
 };
 use crate::{
     core::{
-        cstr::{nstd_core_cstr_as_ptr, nstd_core_cstr_get_null},
+        cstr::{nstd_core_cstr_as_ptr, nstd_core_cstr_get_null, NSTDCStr},
         optional::NSTDOptional,
-        str::{nstd_core_str_as_cstr, NSTDStr},
     },
     cstring::{nstd_cstring_as_ptr, nstd_cstring_from_cstr},
     NSTDAny, NSTDAnyMut, NSTDChar,
@@ -40,7 +39,7 @@ pub type NSTDOptionalSharedLib = NSTDOptional<NSTDSharedLib>;
 ///
 /// # Parameters:
 ///
-/// - `const NSTDStr *path` - A path to the shared library.
+/// - `const NSTDCStr *path` - A path to the shared library.
 ///
 /// # Returns
 ///
@@ -56,23 +55,21 @@ pub type NSTDOptionalSharedLib = NSTDOptional<NSTDSharedLib>;
 ///
 /// - The loaded library may have platform-specific initialization routines ran when it is loaded.
 #[cfg_attr(feature = "clib", no_mangle)]
-pub unsafe extern "C" fn nstd_shared_lib_load(path: &NSTDStr) -> NSTDOptionalSharedLib {
+pub unsafe extern "C" fn nstd_shared_lib_load(path: &NSTDCStr) -> NSTDOptionalSharedLib {
     // Check if `path` is already null terminated.
-    let path = nstd_core_str_as_cstr(path);
-    // Allocate a null byte for `path`.
-    if nstd_core_cstr_get_null(&path).is_null() {
-        let path = nstd_cstring_from_cstr(&path);
+    if nstd_core_cstr_get_null(path).is_null() {
+        // Allocate a null byte for `path`.
+        let path = nstd_cstring_from_cstr(path);
         #[cfg(target_family = "unix")]
         return nstd_os_unix_shared_lib_load(nstd_cstring_as_ptr(&path));
         #[cfg(target_os = "windows")]
         return nstd_os_windows_shared_lib_load(nstd_cstring_as_ptr(&path));
-    }
-    // Use the already null terminated `path`.
-    else {
+    } else {
+        // Use the already null terminated `path`.
         #[cfg(target_family = "unix")]
-        return nstd_os_unix_shared_lib_load(nstd_core_cstr_as_ptr(&path));
+        return nstd_os_unix_shared_lib_load(nstd_core_cstr_as_ptr(path));
         #[cfg(target_os = "windows")]
-        return nstd_os_windows_shared_lib_load(nstd_core_cstr_as_ptr(&path));
+        return nstd_os_windows_shared_lib_load(nstd_core_cstr_as_ptr(path));
     }
 }